requirements.txt
aiohttp==1.0.5
async-timeout==1.2.1
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
discord.py==0.16.8
gitdb2==2.0.2
GitPython==2.1.5
idna==2.5
multidict==3.1.3
pycparser==2.18
PyNaCl==1.0.1
requests==2.18.1
six==1.10.0
smmap2==2.0.3
urllib3==1.21.1
virtualenv==15.1.0
websockets==3.3
youtube-dl==2017.7.15